reading_notes_ 201_Class 04
The biggest asset when you are troubleshooting is your customer. They will be about to provide you key information that could help identify what the issues are and from the information you gather will be about to determine what avenue you should explore and where you should start your troubleshooting . Documentation is also a very vital skill that IT professional must master. There's so much value in complete and accurate documentation, not only for the customer you are servicing but also to any future encounter with similar issues with similar characteristics. The thing that perceived as something not worthy of you exploring because it too basic and simple for that to be the issue, could be the very thing that's creating the issue. Its important to be thorough in how assess a situation and the information that's being provided.